requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2==2.9.9
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.0.0
black==24.4.2
